// src/types/data.ts
export interface DataUsageRecord {
  msisdn: string;
  current_billing_status: string;
  data_usage_raw_total: number;
  apn_name: string;
  event_date: string;
  is_use_internet: string;
  latitude: number;
  longitude: number;
  provinsi: string;
  kabupaten: string;
  kecamatan: string;
  kelurahan: string;
  area: string;
  region: string;
}